recognition/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.5
idna==3.3
imageio==2.13.5
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
lmdb==1.2.1
MarkupSafe==2.0.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
natsort==8.0.2
networkx==2.6.3
numpy==1.21.2
olefile==0.46
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.3.5
pdf2image==1.16.0
pdfkit==1.0.0
Pillow==8.4.0
pip==21.2.4
pyparsing==3.0.6
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
scikit-image==0.19.1
scipy==1.7.3
seaborn==0.11.2
setuptools==58.0.4
six==1.16.0
thefuzz==0.19.0
tifffile==2021.11.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wheel==0.37.0
